requirements.txt
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
eventlet==0.29.1
Flask==1.1.2
Flask-SocketIO==4.3.1
Flask-Sockets==0.2.1
gevent==20.9.0
gevent-websocket==0.10.1
greenlet==0.4.17
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
python-engineio==3.13.2
python-socketio==4.6.0
requests==2.24.0
six==1.15.0
socketIO-client==0.7.2
urllib3==1.25.11
websocket-client==0.57.0
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.1.2
